+++ frontend/src/services/Vault/IVaultService.ts
@@ -1,17 +1,23 @@
 import { storage } from "../../../wailsjs/go/models";
 
 export interface IVaultService {
-    Reshare(
+    reshare(
         vault: any,
         sessionID: any,
         hexEncryptionKey: any,
         serverURL: any): Promise<storage.Vault | undefined>;
 
-    StartKeygen(
+    startKeygen(
         vault: any,
         sessionID: any,
         hexEncryptionKey: any,
         serverURL: any
     ): Promise<storage.Vault | undefined>;
+
+    importVault(buffer: Buffer): Promise<void>;
+
+    encryptVault(passwd: string, vault: Buffer): Buffer;
+
+    decryptVault(passwd: string, vault: Buffer): Buffer;
 
 }
